+++ aquadoggo/src/db/stores/test_utils.rs
@@ -16,7 +16,7 @@ use p2panda_rs::storage_provider::traits::StorageProvider;
 use p2panda_rs::test_utils::constants::{DEFAULT_HASH, DEFAULT_PRIVATE_KEY, TEST_SCHEMA_ID};
 
 use crate::db::provider::SqlStorage;
-use crate::rpc::{EntryArgsRequest, PublishEntryRequest};
+use crate::graphql::client::{EntryArgsRequest, PublishEntryRequest};
 use crate::test_helpers::initialize_db;
 
 pub fn test_operation() -> Operation {
@@ -129,7 +129,7 @@ pub async fn test_db(no_of_entries: usize) -> SqlStorage {
             .await
             .unwrap();
 
-        let backlink = next_entry_args.entry_hash_backlink.clone().unwrap();
+        let backlink = next_entry_args.backlink.clone().unwrap();
 
         // Construct the next UPDATE operation, we use the backlink hash in the prev_op vector
         let update_operation =
@@ -138,8 +138,8 @@ pub async fn test_db(no_of_entries: usize) -> SqlStorage {
         let update_entry = Entry::new(
             &next_entry_args.log_id,
             Some(&update_operation),
-            next_entry_args.entry_hash_skiplink.as_ref(),
-            next_entry_args.entry_hash_backlink.as_ref(),
+            next_entry_args.skiplink.as_ref(),
+            next_entry_args.backlink.as_ref(),
             &next_entry_args.seq_num,
         )
         .unwrap();