+++ aquadoggo/src/graphql/client/mod.rs
@@ -2,6 +2,9 @@
 
 mod request;
 mod response;
+mod root;
+pub(crate) mod u64_string;
 
 pub use request::{EntryArgsRequest, PublishEntryRequest};
 pub use response::{EntryArgsResponse, PublishEntryResponse};
+pub use root::ClientRoot;