+++ aquadoggo/src/graphql/mod.rs
@@ -1,7 +1,8 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 mod api;
+pub mod client;
 mod schema;
 
 pub use api::{handle_graphql_playground, handle_graphql_query};
-pub use schema::{build_static_schema, StaticSchema};
+pub use schema::{build_root_schema, RootSchema};